templates: launch qemu/kvm or container
Revamp the templates so that we launch containers by default and
qemu/kvm if the node is type "qemu". Renamed Config.minicccfs to
Config.default_minicccfs. Added Configs for default_disks or
default_kernel/default_initrd.

Remove S81minirouter to set the filesystem. This should probably be done
upstream (possibly in annotate?).

Expose a lot of the `vm config` API as parameters.

Only works with Go 1.11+.